# src/controllers/dtos/alerts.ts
// Auto-generated by scripts/createDtos.js. Do not edit directly.

export type AlertDtoTypeType = 'EMAIL' | 'SMS' | 'WEBHOOK';

export interface AlertDto {
  id: string;
  domain: string | null;
  type: AlertDtoTypeType;
}

export type UpdateAlertDtoTypeType = 'EMAIL' | 'SMS' | 'WEBHOOK';

export interface UpdateAlertDto {
  id?: string;
  domain?: string | null;
  type?: UpdateAlertDtoTypeType;
}